feat(ws): reuse the shared http.Hijacker upgrade path on the Wing transport
